+++ src/main/java/com/att/tdp/bisbis10/repository/CuisineRepository.java
@@ -3,5 +3,5 @@ package com.att.tdp.bisbis10.repository;
 import com.att.tdp.bisbis10.entity.Cuisine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CuisineRepository extends JpaRepository<Cuisine,Long> {
+public interface CuisineRepository extends JpaRepository<Cuisine,Integer> {
 }
