+++ src/main/java/com/att/tdp/bisbis10/service/DishService.java
@@ -21,7 +21,7 @@ public class DishService {
         dishRepository.save(dish);
     }
 
-    public void deleteDish(Long id){
+    public void deleteDish(Integer id){
         dishRepository.deleteById(id);
     }
